Flyttet UtenlandskID til egen mappe for bedre lesbarhet og forvaltning.
